feat(line filters): move line filters to index scene
